# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type ReplyTone = typeof ReplyTone[keyof typeof ReplyTone];


export const ReplyTone = {
  romantic: 'romantic',
  reassuring: 'reassuring',
  sweet: 'sweet',
  calm: 'calm',
  flirty: 'flirty',
  playful: 'playful',
  natural: 'natural',
  apologetic: 'apologetic',
  confident: 'confident',
  short: 'short',
} as const;

export type ConversationMessageSpeaker = typeof ConversationMessageSpeaker[keyof typeof ConversationMessageSpeaker];


export const ConversationMessageSpeaker = {
  me: 'me',
  them: 'them',
  unknown: 'unknown',
} as const;

export interface ConversationMessage {
  id: string;
  speaker: ConversationMessageSpeaker;
  content: string;
  sequence: number;
  createdAt: string;
}

export interface ReplySuggestion {
  id: string;
  tone: ReplyTone;
  text: string;
  createdAt: string;
}

export interface ConversationSummary {
  id: string;
  title: string;
  /** @nullable */
  preview?: string | null;
  messageCount: number;
  isArchived: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface AnalysisResult {
  id: string;
  conversationId: string;
  summary: string;
  tone: string;
  relationshipContext: string;
  intent: string;
  needsClarification: boolean;
  messages: ConversationMessage[];
  suggestions: ReplySuggestion[];
  createdAt: string;
}

export interface UploadedScreenshot {
  id: string;
  /** @nullable */
  conversationId?: string | null;
  fileName: string;
  contentType: string;
  size: number;
  objectPath: string;
  createdAt: string;
}

export type ConversationDetail = ConversationSummary & ({
  messages: ConversationMessage[];
  analysis: AnalysisResult | null;
  screenshot: UploadedScreenshot | null;
});

export interface ConversationInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  title: string;
}

export interface ConversationUpdate {
  /**
     * @minLength 1
     * @maxLength 120
     */
  title?: string;
  isArchived?: boolean;
}

export interface AnalyzeInput {
  /** @maxLength 16000 */
  conversationText?: string;
  /** @maxLength 4000 */
  latestMessage?: string;
  /** @nullable */
  screenshotId?: string | null;
  /** @maxLength 1000 */
  instruction?: string;
  tone?: ReplyTone;
  includeEmojis?: boolean;
  retainScreenshot?: boolean;
}

export type AnalysisMessageSpeaker = typeof AnalysisMessageSpeaker[keyof typeof AnalysisMessageSpeaker];


export const AnalysisMessageSpeaker = {
  me: 'me',
  them: 'them',
  unknown: 'unknown',
} as const;

export interface AnalysisMessage {
  speaker: AnalysisMessageSpeaker;
  content: string;
  sequence: number;
}

export interface RewriteInput {
  /**
     * @minLength 1
     * @maxLength 200
     */
  instruction: string;
  includeEmojis: boolean;
}

export type UploadRequestContentType = typeof UploadRequestContentType[keyof typeof UploadRequestContentType];


export const UploadRequestContentType = {
  'image/jpeg': 'image/jpeg',
  'image/png': 'image/png',
  'image/webp': 'image/webp',
} as const;

export interface UploadRequest {
  /**
     * @minLength 1
     * @maxLength 255
     */
  name: string;
  /**
     * @minimum 1
     * @maximum 10485760
     */
  size: number;
  contentType: UploadRequestContentType;
}

export interface UploadUrl {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadRequest;
}

export type UploadInputContentType = typeof UploadInputContentType[keyof typeof UploadInputContentType];


export const UploadInputContentType = {
  'image/jpeg': 'image/jpeg',
  'image/png': 'image/png',
  'image/webp': 'image/webp',
} as const;

export interface UploadInput {
  /**
     * @minLength 1
     * @maxLength 512
     */
  objectPath: string;
  /**
     * @minLength 1
     * @maxLength 255
     */
  fileName: string;
  contentType: UploadInputContentType;
  /**
     * @minimum 1
     * @maximum 10485760
     */
  size: number;
  /** @nullable */
  conversationId?: string | null;
}

export type UserPreferenceAppearance = typeof UserPreferenceAppearance[keyof typeof UserPreferenceAppearance];


export const UserPreferenceAppearance = {
  system: 'system',
  light: 'light',
  dark: 'dark',
} as const;

export interface UserPreference {
  defaultTone: ReplyTone;
  includeEmojis: boolean;
  appearance: UserPreferenceAppearance;
}

export type PreferenceInputAppearance = typeof PreferenceInputAppearance[keyof typeof PreferenceInputAppearance];


export const PreferenceInputAppearance = {
  system: 'system',
  light: 'light',
  dark: 'dark',
} as const;

export interface PreferenceInput {
  defaultTone?: ReplyTone;
  includeEmojis?: boolean;
  appearance?: PreferenceInputAppearance;
}

export interface UsageSummary {
  plan: string;
  month: string;
  generations: number;
  imageAnalyses: number;
  storageBytes: number;
  storageLimitBytes: number;
  generationLimit: number;
  imageAnalysisLimit: number;
}

/**
 * Invalid request
 */
export type BadRequestResponse = Error;

/**
 * Sign in is required
 */
export type UnauthorizedResponse = Error;

/**
 * Resource not found
 */
export type NotFoundResponse = Error;

/**
 * Usage limit reached or requests are arriving too quickly
 */
export type RateLimitedResponse = Error;

/**
 * The AI provider could not complete the request
 */
export type UpstreamErrorResponse = Error;