# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [uploads]
      summary: Request a private presigned URL for a screenshot upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadRequest"
      responses:
        "200":
          description: Presigned upload details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrl"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /uploads:
    post:
      operationId: registerUpload
      tags: [uploads]
      summary: Register ownership and metadata after direct screenshot upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadInput"
      responses:
        "201":
          description: Registered private screenshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadedScreenshot"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /uploads/{id}:
    parameters:
      - $ref: "#/components/parameters/UploadId"
    delete:
      operationId: deleteUpload
      tags: [uploads]
      summary: Delete an owned private screenshot
      responses:
        "204":
          description: Screenshot deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    clerkSession:
      type: apiKey
      in: cookie
      name: __session
  parameters:
...
    UploadId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Sign in is required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    UploadRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        size:
          type: integer
          minimum: 1
          maximum: 10485760
        contentType:
          type: string
          enum: [image/jpeg, image/png, image/webp]
      required: [name, size, contentType]
    UploadUrl:
      type: object
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          $ref: "#/components/schemas/UploadRequest"
      required: [uploadURL, objectPath]
    UploadInput:
      type: object
      properties:
        objectPath:
          type: string
          minLength: 1
          maxLength: 512
        fileName:
          type: string
          minLength: 1
          maxLength: 255
        contentType:
          type: string
          enum: [image/jpeg, image/png, image/webp]
        size:
          type: integer
          minimum: 1
          maximum: 10485760
        conversationId:
          type: ["string", "null"]
          format: uuid
      required: [objectPath, fileName, contentType, size]
    UploadedScreenshot:
      type: object
      properties:
        id:
          type: string
          format: uuid
        conversationId:
          type: ["string", "null"]
          format: uuid
        fileName:
          type: string
        contentType:
          type: string
        size:
          type: integer
        objectPath:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, fileName, contentType, size, objectPath, createdAt]
...